% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{cstyle.comment}
\alias{cstyle.comment}
\title{Apply C/C++ style comments}
\usage{
cstyle.comment()
}
\description{
Apply C/C++ style comments
}
\details{
If all lines 'touched' by the cursor or selection start with a
C/C++ style comment, then the method will remove it. Otherwise, it will add the
C/C++ style comment. In the case there is multiple cursors/selection only
the first one (the primary one) will be considered.
}
\author{
Thibault Vatter
}
