% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{ls_objects}
\alias{ls_objects}
\alias{lsos}
\alias{lsos}
\title{Improved list of objects}
\usage{
ls_objects(pos = 1, pattern, order.by, decreasing = FALSE, head = FALSE,
  n = 5)

lsos(..., n = 10)
}
\arguments{
\item{pos}{where to look for (see \code{\link{get}} and \code{\link{ls}})}

\item{pattern}{an optional regular expression (see \code{\link{ls}})}

\item{order.by}{one of \code{c("Type", "Size", "PrettySize", "Rows", "Columns")}}

\item{decreasing}{\code{logical} (for \code{order.by})}

\item{head}{returns the first part of the list}

\item{n}{number of objects if \code{head=TRUE}}
}
\description{
Improved list of objects
}
\author{
http://stackoverflow.com/questions/1358003/tricks-to-manage-the-available-memory-in-an-r-session
}
