% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{lsp}
\alias{lsp}
\title{List functions in a package}
\usage{
lsp(package, all.names = FALSE, pattern)
}
\arguments{
\item{package}{the package name}

\item{all.names}{a logical value (see \code{\link{ls}}). If \code{TRUE}, all
object names are returned. If \code{FALSE}, names which begin with a \code{.} are
omitted.}

\item{pattern}{an optional regular expression (see \code{\link{ls}})}
}
\description{
List functions in a package
}
\author{
Thibault Vatter
}
