% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{roxygen.comment}
\alias{roxygen.comment}
\title{Return a source code with roxygen comments}
\usage{
roxygen.comment(input, start = 1, output = input)
}
\arguments{
\item{input}{input filename.}

\item{start}{starting character in each line.}

\item{output}{output filename.}
}
\description{
Return a source code with roxygen comments
}
\author{
Thibault Vatter
}
