% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{sourceDir}
\alias{sourceDir}
\title{Soure a directory}
\usage{
sourceDir(path = getwd(), trace = TRUE, recursive = TRUE, ...)
}
\arguments{
\item{path}{directory to source}

\item{trace}{print informations}

\item{recursive}{see \code{\link{list.files}}}

\item{...}{additional arguments passed to \code{\link{source}}}
}
\description{
Soure a directory
}
\author{
Thibault Vatter
}
