% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{roxygen.comment}
\alias{roxygen.comment}
\title{Apply roxygen comments}
\usage{
roxygen.comment()
}
\description{
Apply roxygen comments
}
\details{
If all lines 'touched' by the cursor or selection start with a
roxygen2 comment, then the method will remove it. Otherwise, it will add the
roxygen2 comment. In the case there is multiple cursors/selection only
the first one (the primary one) will be considered.
}
\author{
Thibault Vatter
}
